Profile edit, Profile display and Search
Displays list of tutors to choose from with their name and also has a search function, subjects and price. After clicking on the name link the user is also taken to the unique tutor information which contains a bio and a button leading to a chat to book an appointment. The profile edit page is also here enabling the user to change his/her profile information.